Patch YAML loader only within context
Fixes #
